+++ __tests__/client/showAvailableVideos.test.js
@@ -1,8 +1,8 @@
 /**
  * @jest-environment jsdom
  */
-const showAvailableVideos = require("./showAvailableVideos");  
-const basic = require("./basics");  
+const showAvailableVideos = require("../../client/scripts/showAvailableVideos");  
+const basic = require("../../client/scripts/basics");  
 const jsdom = require("jsdom");
 const { JSDOM } = jsdom;
 const dom = new JSDOM();
